repr: split `explain` items in `text`/`json`/`dot` submodules
