+++ src/adapter/src/explain/optimizer_trace.rs
@@ -13,10 +13,9 @@ use std::fmt::Debug;
 
 use mz_compute_client::{plan::Plan, types::dataflows::DataflowDescription};
 use mz_expr::{MirRelationExpr, MirScalarExpr, OptimizedMirRelationExpr, RowSetFinishing};
-use mz_repr::explain::{
-    text_string, DisplayText, Explain, ExplainConfig, ExplainError, ExplainFormat, PlanTrace,
-    TraceEntry,
-};
+use mz_repr::explain::text::{text_string, DisplayText};
+use mz_repr::explain::tracing::{PlanTrace, TraceEntry};
+use mz_repr::explain::{Explain, ExplainConfig, ExplainError, ExplainFormat};
 use mz_sql::plan::{HirRelationExpr, HirScalarExpr};
 use tracing::dispatcher::{self, with_default};
 use tracing_subscriber::prelude::*;
